Add Build Module
This module handles building packages from source inside of a build container.

Additionally it runs Lintian on the results.

The packages are copied into packages/[library]-[version], and a build log is placed in there as well

It throws a BuildError if the build returns nonzero

It throws an LintianError if Lintian detects errors (not warnings) but still copies out the packages

The lintian output is at the end of the log if there was no build error (it is not run if there is a build error)