Hooks are now called from xshop_test.py
XShop calls the run() method automatically in test/xshop_test.py.

Here, the user is passed an object which allows hooks to be run by calling run() and
passing the name of the container and the name of the function to call (any function inside of xshop_test)

The return codes and standard outputs of each hook are returned to XShop.

in xshop_test.py:

def run(H):
	H.run('target','run_exploit')

def run_exploit():

The three examples have been updated to use this, and an example of the XShop
API is included in Hearbleed/run