XSHOP Pull
There are three new fields in config.yaml

First, `source` list for listing source urls,

these can be a just a list or strings, or if a string has a bracket list
in it, this is expanded to multiple strings. For example:

http://example.com/library-{version1,version2}.tar.gz

is expanded to:

http://example.com/library-version1.tar.gz
http://example.com/library-version2.tar.gz

Second, `public_keys` is a list of public keys associated with files.

XShop does not use this, other than to notify the user when they pull.

XShop still expects the user to manually install keys.

Third, `notes`

This is a string which is printed after xshop pull to notify the user of
any specifics about this project. If you do this:

notes: |
	line 1
	line 2

Then the newlines are preserved and you can print multiple lines to the
user.

For users, if these values are present, running `xshop pull` in the root
of a project will download all source files, and attempt to verify with
pgp signature, md5, or sha1 in that order.

If all hashes are stored in a CHECKSUMS file, then it will attempt to
verify against these using MD5.

For pgp keys, the keys covering all of the files should be listed in
config.yaml and printed during xshop pull. The user should manually
import these keys (we dont want to trust them implicitly on the users
behalf).