Parsed Data into nice table
Output was converted to real JSON, was a printed object. I'm not sure about modifying XShop to output json, it was throwing errors before

The table is a colored text, view it with `cat Table`.

It confirms that patch 27 fixed everything.

Worth noting, CVE-2014-7187 and the unamed CVE checked for did not affect any versions.
I manually checked a number of them, with different tests found online, and it definitely doesnt have any affect.
I'm not sure what is going on there, but supposedly they were patched in 27.